+++ storage/src/main/scala/uk/ac/wellcome/storage/dynamo/DynamoVersionedDao.scala
@@ -2,7 +2,7 @@ package uk.ac.wellcome.storage.dynamo
 
 import com.amazonaws.services.dynamodbv2.model.ProvisionedThroughputExceededException
 import grizzled.slf4j.Logging
-import uk.ac.wellcome.storage.BetterVersionedDao
+import uk.ac.wellcome.storage.VersionedDao
 import uk.ac.wellcome.storage.type_classes.{VersionGetter, VersionUpdater}
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -14,7 +14,7 @@ class DynamoVersionedDao[T](
   val versionGetter: VersionGetter[T],
   val versionUpdater: VersionUpdater[T])
     extends Logging
-    with BetterVersionedDao[T] {
+    with VersionedDao[T] {
 
   @deprecated("Use put() instead of updateRecord()", since = "2019-05-10")
   def updateRecord[R](record: T): Future[T] =
