+++ storage/src/main/scala/uk/ac/wellcome/storage/memory/MemoryVersionedDao.scala
@@ -1,6 +1,6 @@
 package uk.ac.wellcome.storage.memory
 
-import uk.ac.wellcome.storage.BetterVersionedDao
+import uk.ac.wellcome.storage.VersionedDao
 import uk.ac.wellcome.storage.type_classes.{VersionGetter, VersionUpdater}
 
 class MemoryVersionedDao[T](
@@ -9,4 +9,4 @@ class MemoryVersionedDao[T](
   implicit
   val versionGetter: VersionGetter[T],
   val versionUpdater: VersionUpdater[T]
-) extends BetterVersionedDao[T]
+) extends VersionedDao[T]
